Update the AV::Template extension to support reloading in Rails 6+
Rely on #encode! as the hook in which to update the view source and
the handler.

Use #compile! as the hook in which to check if the last compilation
has expired after the overrides were updated.

Detect syntax and apply overrides in the template helper using the
same, more explicit, technique.